# app/Http/Requests/StoreReservationRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Validation\Rule;

/**
 * @OA\Schema(
 *     schema="StoreReservationRequest",
 *     type="object",
 *     title="Store Reservation Request",
 *     required={"user_id", "start_time", "end_time"},
 *     @OA\Property(
 *         property="user_id",
 *         type="integer",
 *         description="ID of the user"
 *     ),
 *     @OA\Property(
 *         property="start_time",
 *         type="string",
 *         format="date-time",
 *         description="Start time of the reservation"
 *     ),
 *     @OA\Property(
 *         property="end_time",
 *         type="string",
 *         format="date-time",
 *         description="End time of the reservation"
 *     )
 * )
 */
class StoreReservationRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'user_id' => 'required|exists:users,id',
            'start_time' => [
                'required',
                'date_format:Y-m-d H:i:s',
                Rule::unique('reservations', 'start_time'),
                'before:end_time'
            ],
            'end_time' => [
                'required',
                'date_format:Y-m-d H:i:s',
                Rule::unique('reservations', 'end_time'),
                'after:start_time'
            ],
        ];
    }

    public function messages()
    {
        return [
            'start_time.required' => 'The start time is required.',
            'start_time.date_format' => 'The start time must be in the format Y-m-d H:i:s.',
            'start_time.unique' => 'The start time must be unique.',
            'start_time.before' => 'The start time must be before the end time.',
            'end_time.required' => 'The end time is required.',
            'end_time.date_format' => 'The end time must be in the format Y-m-d H:i:s.',
            'end_time.unique' => 'The end time must be unique.',
            'end_time.after' => 'The end time must be after the start time.',
        ];
    }
}