# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SportReserve",
        "contact": {
            "email": "derex@outlook.com.br"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-sport-reserve.juvhost.com"
        }
    ],
    "paths": {
...
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get list of reservations",
                "description": "Returns list of reservations",
                "operationId": "getReservationsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Store a new reservation",
                "description": "Stores a new reservation and returns the reservation data",
                "operationId": "storeReservation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get reservation information",
                "description": "Returns reservation data",
                "operationId": "getReservationById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete an existing reservation",
                "description": "Deletes an existing reservation and returns success message",
                "operationId": "deleteReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update an existing reservation",
                "description": "Updates an existing reservation and returns the updated data",
                "operationId": "updateReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreReservationRequest": {
                "title": "Store Reservation Request",
                "required": [
                    "user_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "user_id": {
                        "description": "ID of the user",
                        "type": "integer"
                    },
                    "start_time": {
                        "description": "Start time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "description": "End time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "UpdateReservationRequest": {
                "title": "Update Reservation Request",
                "properties": {
                    "user_id": {
                        "description": "ID of the user",
                        "type": "integer"
                    },
                    "start_time": {
                        "description": "Start time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "description": "End time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Reservation": {
                "title": "Reservation",
                "required": [
                    "user_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the reservation",
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "description": "ID of the user",
                        "type": "integer"
                    },
                    "start_time": {
                        "description": "Start time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_time": {
                        "description": "End time of the reservation",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    }
}